fix: handle unrecognized language codes gracefully
Previously, the language list would fail to load if the tessdata
directory contained any files with unrecognized language codes, as
sorting would raise an exception for unknown languages.

The check for valid codes, which was already implemented in
`get_downloaded_languages()`, has been moved to be done earlier in
`get_downloaded_codes()` to prevent these errors. Now, any unrecognized
codes are ignored, with a warning message logged to the console.