chore: remove invalid `–psm 0` option from tessdata_config
The `–psm 0` option was ignored by Tesseract because it used an en-dash
instead of a hyphen, and only a single dash instead of the required
double dash.

Using the correct `--psm 0` is also not appropriate, as it triggers
"Orientation and script detection (OSD) only", which does return any
text and causes frog to fail.

As a result, the behavior remains unchanged and continues to default to
"Fully automatic page segmentation, but no OSD".